// src/core-utils/interfaces/results/subjectsMarks.interface.ts
export interface ISubjectMarks {
    subject: string;
    int1_max: number;
    int1: number;
    int2_max: number;
    int2: number;
    assn1_max: number;
    assn1: number;
    assn2_max: number;
    assn2: number;
    assn3_max: number;
    assn3: number;
    quiz1_max: number;
    quiz1: number;
    quiz2_max: number;
    quiz2: number;
    quiz3_max: number;
    quiz3: number;
    sess_max: number;
    sess: number;
    ext_grade: string;
}